# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MrAlgorithm API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/goods/{good}/parameters/plati": {
            "post": {
                "tags": [
                    "Goods"
                ],
                "summary": "Publish Parameter model records to a Plati product",
                "description": "Requires good_manage_plati_parameters. Reads Parameter records and options directly; no request body is required. Parameter titles are sent to Plati for both en-US and ru-RU using the same local title. Existing Plati mappings are reused and missing mappings are saved from IDs returned by Plati.",
                "operationId": "publishGoodParametersToPlati",
                "parameters": [
                    {
                        "name": "good",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parameters published successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Good parameters published to Plati successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "parameter_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "marketplace_parameter_id": {
                                                        "type": "integer",
                                                        "example": 602080
                                                    },
                                                    "marketplace_option_ids": {
                                                        "type": "object",
                                                        "example": {
                                                            "1": 6020801
                                                        },
                                                        "additionalProperties": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "created": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Missing good_manage_plati_parameters permission"
                    },
                    "404": {
                        "description": "Good not found"
                    },
                    "422": {
                        "description": "Good is unpublished, no parameters exist, or Plati rejected a parameter"
                    },
                    "502": {
                        "description": "Plati connection or response failure"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token returned by POST /api/login.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}